Refactored

+++ src/controllers/main_controller.py
@@ -3,38 +3,38 @@ from typing import Optional
 import i18n
 from openpyxl.utils.exceptions import InvalidFileException
 
-from handlers import session_handlers
-from data.learning_data import LearningData
-from GUIs.list_gui import WordList
-from GUIs.main_gui import MainGui
-from data.repository import RepositoryProtocol
-from controllers.session_controller import SessionController
-from GUIs.session_gui import SessionGui
-from settings import UserSettings, settings
+from src.handlers import session_handlers
+from src.data.learning_data import LearningData
+from src.GUIs.list_gui import WordList
+from src.GUIs.main_gui import MainGui
+from src.data.repository import RepositoryProtocol
+from src.controllers.session_controller import SessionController
+from src.GUIs.session_gui import SessionGui
+from src.settings import AppSettings, SessionSettings
 
 
 class MainController:
-    def __init__(self, gui: MainGui, repository: RepositoryProtocol, handlers):
+    def __init__(self, gui: MainGui, repository: RepositoryProtocol, handlers, settings: AppSettings):
         self.session_controller = None
         self.gui = gui
         self.repository = repository
         self.learning_data: LearningData = None
         self.handlers = handlers
+        self.settings = settings
 
     def start_app(self, name: Optional[str] = None):
         if name not in self.repository.available_dictionaries:
             name = self.repository.available_dictionaries[0]
         self.learning_data = LearningData(name=name, **self.repository.get_dictionary(name))
-
-        self.gui.configure_ui(self.handlers.get_user_settings(settings.USER_SETTINGS_PATH))
         self.update_widgets()
         self.gui.set_commands(self)
         self.gui.bind_events(self)
         self.change_word_per_session()
+
         self.gui.mainloop()
 
     def exit_app(self):
-        self.handlers.save_user_settings(self.gui.settings, settings.USER_SETTINGS_PATH)
+        self.handlers.save_user_settings(self.gui.settings, self.settings.USER_SETTINGS_PATH)
         self.repository.update_dictionary(
             name=self.learning_data.name,
             dictionary=self.learning_data.dictionary,
@@ -48,7 +48,8 @@ class MainController:
             repository=self.repository,
             learning_data=self.learning_data,
             handlers=session_handlers,
-            settings=UserSettings(**self.gui.settings)
+            session_settings=SessionSettings(**self.gui.session_settings),
+            app_settings=self.settings
         )
         self.session_controller.start_session()
 
@@ -58,6 +59,7 @@ class MainController:
             data = self.handlers.get_data(
                 name=self.gui.new_dictionary_name,
                 taken_names=self.repository.available_dictionaries,
+                settings=self.settings
             )
             # add data to db
             self.repository.add_dictionary(**data)
@@ -70,6 +72,8 @@ class MainController:
     def remove_dictionary(self, name: str):
         if len(self.repository.available_dictionaries) == 1:
             return self.gui.show_error(title="error", message="forbidden")
+        if not self.gui.get_user_confirmation(i18n.t("remove_dict_lb")):
+            return
         self.repository.delete_dictionary(name)
         # restart the app
         self.start_app()
@@ -106,8 +110,12 @@ class MainController:
 
     def show_word_list(self):
         word_list = WordList()
-        word_list.fill_in_tree(self.learning_data.dictionary, self.learning_data.scores, settings.MAX_PHRASE_LENGTH//2)
-        word_list.configure_style(self.gui.appearance_mode)
+        word_list.fill_in_tree(self.learning_data.dictionary, self.learning_data.scores, self.settings.MAX_PHRASE_LENGTH//2)
+        word_list.configure_style(self.gui.appearance_mode, self.settings)
+
+    def change_app_language(self, language: str):
+        i18n.set("locale", self.settings.LOCALE[language])
+        self.gui.set_texts()
 
     def filter_remove_word_combobox(self, event: Event):
         if not hasattr(self.gui, "filter_remove_word_combobox"):
